Update structura script player + minimal AI

+++ Assets/Scripts/Enemy.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health;
+    public float health;
     public float speed;
     public float deathTime;
     public Animator animator;
@@ -36,7 +36,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
         Debug.Log("A LUAT DAMAGE !!!");
